Marker changes and agent integration
