+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/agents/SBO/TypePriority.java
@@ -1,9 +1,5 @@
 package com.nscharrenberg.um.multiagentsurveillance.agents.SBO;
 
-import com.nscharrenberg.um.multiagentsurveillance.headless.models.*;
-import com.nscharrenberg.um.multiagentsurveillance.headless.models.Marker;
-import com.nscharrenberg.um.multiagentsurveillance.headless.models.MarkerSmell;
-
 public enum TypePriority {
     SoundWave(3),
     MarkerDeadEnd(-5),
