+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/headless/models/Items/SoundWave.java
@@ -1,24 +1,21 @@
-package com.nscharrenberg.um.multiagentsurveillance.headless.models;
+package com.nscharrenberg.um.multiagentsurveillance.headless.models.Items;
 
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Angle.AdvancedAngle;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Items.Item;
-import com.nscharrenberg.um.multiagentsurveillance.headless.models.Items.ItemType;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Map.Tile;
-import com.nscharrenberg.um.multiagentsurveillance.headless.models.Marker.MarkerType;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Player.Player;
 
 
-public class MarkerSmell extends Marker {
+public class SoundWave extends Item {
+    private int strength; // Sound strength
+    private AdvancedAngle direction; // Sound direction (where it is coming from)
+    private Player owner;
 
-    private int strength;
-    private AdvancedAngle direction;
-
-
-    // TODO: We need some way of distinguishing markers (and add it to itemtype)
-    public MarkerSmell(Tile tile, Marker.MarkerType marking, int strength, AdvancedAngle direction, Player player, int duration) {
-        super(marking, tile, player, duration);
+    public SoundWave(Tile tile, int strength, AdvancedAngle direction, Player owner) {
+        super(tile);
         this.strength = strength;
         this.direction = direction;
+        this.owner = owner;
     }
 
     public int getStrength() {
@@ -36,4 +33,9 @@ public class MarkerSmell extends Marker {
     public void setDirection(AdvancedAngle direction) {
         this.direction = direction;
     }
+
+    public Player getOwner() { return owner; }
+
+    public void setOwner(Player owner) { this.owner = owner; }
+
 }
