+++ src/main/java/com/nscharrenberg/um/multiagentsurveillance/headless/utils/CharacterAudio.java
@@ -3,8 +3,7 @@ package com.nscharrenberg.um.multiagentsurveillance.headless.utils;
 import com.nscharrenberg.um.multiagentsurveillance.headless.exceptions.ItemAlreadyOnTileException;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Angle.*;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Map.TileArea;
-import com.nscharrenberg.um.multiagentsurveillance.headless.models.SoundWave;
-import com.nscharrenberg.um.multiagentsurveillance.headless.models.Items.*;
+import com.nscharrenberg.um.multiagentsurveillance.headless.models.Items.SoundWave;
 import com.nscharrenberg.um.multiagentsurveillance.headless.models.Map.*;
 
 import java.util.HashMap;
