+++ src/test/resources/maps/trainingExampleMap.txt
@@ -13,8 +13,8 @@ timeStep = 0.1
 targetArea = 45 20 49 24
 spawnAreaIntruders = 2 2 8 8
 spawnAreaGuards = 2 2 8 8
-deadendMarker = 22 9
-deadendMarker = 22 19
+deadendMarker = 34 24
+deadendMarker = 34 1
 wall = 0 0 50 0
 wall = 0 0 0 25
 wall = 50 0 50 50
@@ -22,21 +22,21 @@ wall = 0 25 50 25
 wall = 44 19 46 19
 wall = 48 19 50 19
 wall = 44 19 44 25
-wall = 46 12 46 19
+wall = 46 13 46 19
 wall = 46 1 46 10
 wall = 48 1 48 19
 wall = 35 10 45 10
-wall = 35 12 45 12
+wall = 35 13 45 13
 wall = 35 1 35 10
 wall = 33 1 33 16
-wall = 33 18 33 25
-wall = 35 12 35 25
+wall = 33 19 33 25
+wall = 35 13 35 25
 wall = 23 16 33 16
-wall = 23 18 33 18
-wall = 23 18 23 25
+wall = 23 19 33 19
+wall = 23 19 23 25
 wall = 23 1 23 16
-wall = 21 1 21 8
+wall = 21 1 21 7
 wall = 21 10 21 25
 wall = 1 10 21 10
-wall = 10 8 21 8
-wall = 10 1 10 8
+wall = 10 7 21 7
+wall = 10 1 10 7
